WHO ques modified

+++ main_file.py
@@ -30,14 +30,17 @@ solr = pysolr.Solr('http://localhost:8983/solr/final1', timeout=10)
 
 
 def readQuestions(fileName):
-    with open(fileName, 'r') as f:
-        temp = f.read().splitlines()
-    #function call to delete existing answers.json file if it exists:
-    deleteJSONFile()
-    
-    #function call to process questions
-    processQuestions(temp)
-    #return temp
+    if(os.path.isfile(fileName)):
+        with open(fileName, 'r') as f:
+            temp = f.read().splitlines()
+        #function call to delete existing answers.json file if it exists:
+        deleteJSONFile()
+        
+        #function call to process questions
+        processQuestions(temp)
+        #return temp
+    else:
+        print("ERROR: Please provide a valid path for the questions file in the arguments.")
 
 
 
@@ -97,7 +100,7 @@ def processQuestions(input_questions):
         (hypernyms_list:"+hypernyms_list+") OR (hyponyms_list:"+hyponyms_list+") OR (stemmatize_word:"+stemmatize_word+")^10 AND (entities_list:"+entities_list+")^20)"
         
         #print("sentence:"+sentence)
-        #print(query)
+        print(query)
         
         print("Query created for the question")
         #print(query)
@@ -115,36 +118,44 @@ def processQuestions(input_questions):
     
 def getAnswer(query,term1,term2,entities_list):
     results = None
-    results = solr.search(q=query,start=0, rows=1)
+    results = solr.search(q=query,start=0, rows=10)
     if len(results) == 0:
         return "Answer not found", "N.A.", "N.A."
     #print("length of the results", len(results))
-    #counter = 0
+    counter = 0
     for result in results:
+        print(result['sentence'])
         ######## code to check entities
         #counter += 1
-        #ans_entity_list = result['entities_list']
-        #flag = False
-        #for e1 in ans_entity_list:
-            #for e2 in  entities_list:
-                #if((e1 in e2) or (e2 in e1)):
-                    #flag = True
-                    #break
-            #if(flag==True):
-                #break
-        #if(not flag and counter<len(results)-1):
-            #continue
+#        ans_entity_list = result['entities_list']
+#        flag = False
+#        print((ans_entity_list))
+#        print((entities_list))
+#        for e1 in ans_entity_list:
+#            for e2 in  entities_list:
+#                print(e1+":::"+e2)
+#                if((e1 in e2) or (e2 in e1)):
+#                    print("flag made true")
+#                    flag = True
+#                    break
+#            if(flag==True):
+#                break
+#        if(not flag and counter<len(results)-1):
+#            continue
         
             
         
         #print("The title is '{0}','{1}'.".format(result['sentence'],result['name']))
-        doc = nlp(str(result['sentence']))
+        doc = nlp(str(result['sentence'][0]))
         answer = ""
         for X in doc.ents:
             if(X.label_ == term1 or X.label_ == term2):
+                ######### for WHO questions: ensure that that ensure that answer is not same/
+                # as the question's main entity. ex: Who founded Apple? shouldn't return 'Apple'
                 if(term1 == 'PERSON'):
-                    #print(result['sentence'],"::",X.text,"**", entities_list)
-                    if (not (str(X.text) in entities_list)): #return first answer for WHO questions
+                    #print(type(result['sentence'][0]),"::",X.text,"**", entities_list)
+                    if (not (X.text in entities_list)):# filter out the entity in question from the answer
+                        #print("X.text",X.text)
                         answer = X.text
                         break
                 else:
@@ -153,6 +164,10 @@ def getAnswer(query,term1,term2,entities_list):
             #print("Answer is--> ",X.text)
     #if(answer[:-1].isnumeric()):
         #return answer[:-1], result['sentence'][0], result['name'][0]
+        counter += 1
+        if(len(answer)==0 and counter<10):
+            continue
+        
         return answer, result['sentence'][0], result['name'][0]
